Renamed region-classifier scripts
